Add root and flow for app
