+++ src/app/data.js
@@ -1,5 +1,3 @@
-import { Flow } from "../flow/index.js";
-
 export function app$data(data) {
-  return new Flow([data], data, this, this);
+  return this._flow.data(data);
 }
